:hammer: Move AppPathProvider interface to desktopMain for desktop-specific impl
