--- composeApp/src/commonMain/kotlin/com/crosspaste/path/AppPathProvider.kt
@@ -1,16 +1,0 @@
-package com.crosspaste.path
-
-import okio.Path
-
-interface AppPathProvider {
-
-    val userHome: Path
-
-    val pasteAppPath: Path
-
-    val pasteAppJarPath: Path
-
-    val pasteAppExePath: Path
-
-    val pasteUserPath: Path
-}
